Only knock-on after filtering out some random touch screen events.
For some reason or other, our touch screen isn't happy with Android 5.1, and
is occasionally getting probed because it thinks something is wrong. Luckily
when it gets probed, it is not registering as a knock-on event, so we can do
some logic to ensure that is IS a true knock-on event, and wake up only when
a knock-on event is registered, rather than always knocking on when we get to
this particular function.

Change-Id: I4a629eff939f9b22213bdda3ddf221fb82cfd5c1